Flytt hjelpemetode for dsop-meldekort ut av repository
